Relocated collections utils
